add account config in data_aug scripts.

+++ llava/data_aug/reformat_tar.py
@@ -6,16 +6,23 @@ import tarfile
 def main(
     src_tar="~/workspace/sa_000000.tar", 
     src_folder="~/workspace",
-    tgt_folder="~/workspace/dev-123"
+    tgt_folder="~/workspace/dev-123",
+    overwrite=False,
 ):
     
     src_tar_path = osp.expanduser(src_tar)
     src_folder_path = osp.expanduser(src_folder)
     tgt_folder_path = osp.expanduser(tgt_folder)
     rpath = osp.relpath(src_tar_path, src_folder_path)
+    fpath = osp.join(tgt_folder_path, rpath)
+    
+    if osp.exists(fpath) and not overwrite:
+        print("Skipping")
+        return 
+        
+    
     t = tarfile.open(src_tar_path)
 
-    fpath = osp.join(tgt_folder_path, rpath)
     os.makedirs(osp.dirname(fpath), exist_ok=True)
     tdev = tarfile.open(fpath, "w")
